requirements.txt
# ViZDoom Deep RL Ablation Study
# Python 3.12 compatible

# Core RL environment
vizdoom==1.2.4
gymnasium==1.2.3

# Deep learning
torch==2.9.1
torchvision==0.24.1

# Numerical computing
numpy==2.4.0
opencv-python==4.12.0.88

# Experiment tracking
wandb==0.23.1

# Configuration management
hydra-core==1.3.2
omegaconf>=2.3.0

# Visualization
matplotlib==3.10.8

# Development (optional)
# pytest>=8.0.0
# black>=24.0.0
# ruff>=0.1.0
